Opal: fix Windows build
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/opal/trunk@30881 72102866-910b-0410-8b05-ffd578937521

+++ Source/CGDataProvider.m
@@ -180,7 +180,7 @@
   }
   else if (getBytesAtPositionCallback)
   {
-    return getBytesAtPositionCallback(info, buffer, position, count);
+    return getBytesAtPositionCallback(info, buffer, (int)position, count);
   }
   return 0;
 }
